refactor metadata into own package.
