+++ src/dakoda/__init__.py
@@ -1,1 +1,2 @@
-from .corpus import DakodaCorpus as Dakoda
+from .corpus import DakodaCorpus as Dakoda, DakodaDocument
+from .metadata import MetaData
