+++ src/dakoda/metadata/__init__.py
@@ -1,0 +1,1 @@
+from .scheme import MetaData
