--- src/dakoda/types.py
@@ -1,8 +1,0 @@
-T_TOKEN      = "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token"
-T_LEMMA      = "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Lemma"
-T_POS        = "de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS"
-T_SENT       = "de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence"
-T_MORPH      = "de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.morph.Morpheme"
-T_DEP        = "de.tudarmstadt.ukp.dkpro.core.api.syntax.type.dependency.Dependency"
-
-T_META       = "de.tudarmstadt.ukp.dkpro.core.api.metadata.type.MetaDataStringField"
